+++ firmware/controllers/math/engine_math.cpp
@@ -26,14 +26,14 @@
 #include "advance_map.h"
 #include "gppwm_channel.h"
 
-floatms_t getEngineCycleDuration(int rpm) {
+floatms_t getEngineCycleDuration(float rpm) {
 	return getCrankshaftRevolutionTimeMs(rpm) * (getEngineRotationState()->getOperationMode() == TWO_STROKE ? 1 : 2);
 }
 
 /**
  * @return number of milliseconds in one crank shaft revolution
  */
-floatms_t getCrankshaftRevolutionTimeMs(int rpm) {
+floatms_t getCrankshaftRevolutionTimeMs(float rpm) {
 	if (rpm == 0) {
 		return NAN;
 	}
